+++ packages/apple-targets/package.json
@@ -30,7 +30,11 @@
     "config-plugins",
     "apple-targets"
   ],
-  "author": "Evan Bacon",
+  "author": {
+    "email": "baconbrix@gmail.com",
+    "name": "Evan Bacon",
+    "url": "https://evanbacon.dev"
+  },
   "license": "MIT",
   "dependencies": {
     "@bacons/xcode": "1.0.0-alpha.27",
