+++ packages/apple-targets/src/config-plugin.ts
@@ -7,8 +7,8 @@ import type { Config, ConfigFunction } from "./config";
 import { withPodTargetExtension } from "./withPodTargetExtension";
 import withWidget from "./withWidget";
 import { withXcodeProjectBetaBaseMod } from "./withXcparse";
+import { warnOnce } from "./util";
 
-let hasWarned = false;
 export const withTargetsDir: ConfigPlugin<
   {
     appleTeamId?: string;
@@ -20,9 +20,8 @@ export const withTargetsDir: ConfigPlugin<
   const { root = "./targets", match = "*" } = _props || {};
   const projectRoot = config._internal!.projectRoot;
 
-  if (!appleTeamId && !hasWarned) {
-    hasWarned = true;
-    console.warn(
+  if (!appleTeamId) {
+    warnOnce(
       chalk`{yellow [bacons/apple-targets]} Expo config is missing required {cyan ios.appleTeamId} property. Find this in Xcode and add to the Expo Config to correct. iOS builds may fail until this is corrected.`
     );
   }
